+++ Realm/ObjectServerTests/SwiftSyncTestCase.swift
@@ -50,7 +50,7 @@ class SwiftSyncTestCase: RLMSyncTestCase {
             semaphore.signal()
         }
         SyncManager.shared.setSessionCompletionNotifier(basicBlock)
-        let config = Realm.Configuration(syncConfiguration: (user, url))
+        let config = Realm.Configuration(syncConfiguration: SyncConfiguration(user: user, realmURL: url))
         let realm = try Realm(configuration: config)
         // FIXME: Perhaps we should have a reasonable timeout here, instead of allowing bad code to stall forever.
         _ = semaphore.wait(timeout: .distantFuture)
@@ -58,7 +58,7 @@ class SwiftSyncTestCase: RLMSyncTestCase {
     }
 
     func immediatelyOpenRealm(url: URL, user: SyncUser) throws -> Realm {
-        return try Realm(configuration: Realm.Configuration(syncConfiguration: (user, url)))
+        return try Realm(configuration: Realm.Configuration(syncConfiguration: SyncConfiguration(user: user, realmURL: url)))
     }
 
     func synchronouslyLogInUser(for credential: Credential,
@@ -124,7 +124,7 @@ class SwiftSyncTestCase: RLMSyncTestCase {
             dispatch_semaphore_signal(semaphore)
         }
         SyncManager.sharedManager().setSessionCompletionNotifier(basicBlock)
-        let config = Realm.Configuration(syncConfiguration: (user, url))
+        let config = Realm.Configuration(syncConfiguration: SyncConfiguration(user: user, realmURL: url))
         let realm = try Realm(configuration: config)
         // FIXME: Perhaps we should have a reasonable timeout here, instead of allowing bad code to stall forever.
         dispatch_semaphore_wait(semaphore, DISPATCH_TIME_FOREVER)
@@ -132,7 +132,7 @@ class SwiftSyncTestCase: RLMSyncTestCase {
     }
 
     func immediatelyOpenRealm(url: NSURL, user: SyncUser) throws -> Realm {
-        return try Realm(configuration: Realm.Configuration(syncConfiguration: (user, url)))
+        return try Realm(configuration: Realm.Configuration(syncConfiguration: SyncConfiguration(user: user, realmURL: url)))
     }
 
     func synchronouslyLogInUser(for credential: Credential, server url: NSURL) throws -> SyncUser {